+++ releases/0.2.0/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Version 0.2.0 (November 8, 2019)
 
 * Feature: Add document.render() / El.renderPage() to render the entire page right after it is loaded (w00t!)
+* Feature: Provide document.binding / El.pageBinding returning the binded object for page rendering
 * Feature: Support using `.` for interpolating the binding itself (e.g. strings, integers)
 
 ### Version 0.1.1 (November 7, 2019)